AD-HOC fix (pvc): Fix a bug with storage class, add support for matchLabels
Currently there is a problem in which, if a resource already exists,
it's difficult to match it against the appropriate volume. The existing
solution for this is to create a new storage class specifically for a
resource and invoke it with a PVC, however there appears to be a new,
superior mechanism -- matching labels.

This commit introduces support for this (available at least in 1.9+), as
well as fixes a derp with naming in the storage class.